workshops/functions/packages/ynov/dorking/requirements.txt
googlesearch-python
